# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API",
    "version": "1.0.0",
    "description": "API documentation for the blog app"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "summary": "Get all blog posts",
        "responses": {
          "200": {
            "description": "A list of blog posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogPost"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new blog post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog post created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "summary": "Get a blog post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the blog post"
          }
        ],
        "responses": {
          "200": {
            "description": "A single blog post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "put": {
        "summary": "Update a blog post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the blog post"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogPost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog post updated"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a blog post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the blog post"
          }
        ],
        "responses": {
          "204": {
            "description": "Blog post deleted"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlogPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the blog post"
          },
          "title": {
            "type": "string",
            "description": "Title of the blog post"
          },
          "content": {
            "type": "string",
            "description": "Content of the blog post"
          },
          "imageUrl": {
            "type": "string",
            "description": "Upload image url or image file with fieldname `image`"
          }
        },
        "required": ["title", "content"]
      }
    }
  }
}